refactor: 스크롤 방지 훅
